Finish coordinator refactoring
