+++ app/src/device/kotlin/de/digitalService/useID/MainActivity.kt
@@ -17,8 +17,7 @@ import de.digitalService.useID.idCardInterface.IdCardManager
 import de.digitalService.useID.models.NfcAvailability
 import de.digitalService.useID.ui.UseIDApp
 import de.digitalService.useID.ui.coordinators.AppCoordinatorType
-import de.digitalService.useID.ui.coordinators.AppNavigator
-import de.digitalService.useID.ui.coordinators.Navigator
+import de.digitalService.useID.ui.navigation.Navigator
 import de.digitalService.useID.util.NfcAdapterUtil
 import io.sentry.Sentry
 import kotlinx.coroutines.CoroutineScope
@@ -118,15 +117,11 @@ class MainActivity : ComponentActivity() {
         when (intent.action) {
             NfcAdapter.ACTION_TAG_DISCOVERED -> {
                 intent.getParcelableExtra<Tag>(NfcAdapter.EXTRA_TAG)?.let {
-//                    if (appCoordinator.currentlyHandlingNfcTags) {
-                        try {
-                            idCardManager.handleNfcTag(it)
-                        } catch (e: IOException) {
-                            logger.error("IDCardManager failed to handle NFC tag.")
-                        }
-//                    } else {
-//                        logger.debug("Got new NFC tag but app coordinator is not awaiting any. Ignoring.")
-//                    }
+                    try {
+                        idCardManager.handleNfcTag(it)
+                    } catch (e: IOException) {
+                        logger.error("IDCardManager failed to handle NFC tag.")
+                    }
                 }
             }
