+++ app/src/main/kotlin/de/digitalService/useID/ui/coordinators/SubCoordinatorState.kt
@@ -1,0 +1,5 @@
+package de.digitalService.useID.ui.coordinators
+
+enum class SubCoordinatorState {
+    Idle, Active, Finished, Cancelled
+}
